refactor(math): move Savage-Dickey primitives to src/math/savage_dickey/
The 1D integrals and cubic critical-point solver under models/ggm/sd_density_*
are pure math — their APIs take (A, B, s_jj, alpha) doubles with no GGM state
— so they belong alongside cholupdate and cholesky_helpers, not under a model
directory. Moves and renames in one go (git tracks all as renames):

  models/ggm/sd_density_cubic.{h,cpp}          -> math/savage_dickey/cubic_mode.{h,cpp}
  models/ggm/sd_density_l_space.{h,cpp}        -> math/savage_dickey/laplace.{h,cpp}
  models/ggm/sd_density_l_space_quad.{h,cpp}   -> math/savage_dickey/gauss_hermite.{h,cpp}
  models/ggm/sd_density_l_space_sinh.{h,cpp}   -> math/savage_dickey/sinh_midpoint.{h,cpp}

Namespace ggm_sd -> savage_dickey. ggm_model.cpp and log_z_test_interface.cpp
include paths and qualifiers updated; comment references to the old filenames
in both files and in test-sd-{cubic,sinh}.R updated to the new paths.

Behaviour-preserving: sd-baseline-snapshot, sd-cubic, and sd-sinh tests all
pass against the alpha=1 and alpha=3 baselines committed pre-refactor.
